Bootloader: never interrupt spm-executing boot_* function.
All boot_* functions executing SPM instruction need to be called in
ATOMIC_BLOCK - interrupts need to be disabled during execution of the function,
because the following needs to be met [ATmega128A datasheet]:
"To execute page erase, ... write “X0000011” to SPMCSR and execute SPM
**within four clock cycles** after writing SPMCSR."
This is a real issue on in MTB-UNI-2-AVR-FW with baudrate 230 400 bd/s.
Random words in flash were not programmed after an attempt to reprogram
the flash.